move files to outer shared scope
